tui: Move action dispatch table
Interactive mode now delegates individual action implementations to focused
TUI modules, but the dispatcher table still lives beside the session loop.

The loop cannot stay small while it owns registration, hunk requirements,
escape handling, and action-specific adapters.

This commit addresses that by moving the action handler table and dispatch
routine into tui.action_dispatch and calling it from the loop.

Subsequent commits will retarget dispatcher tests and update adapter
boundary guards.
